Fix issue #12 (Fails to parse Wallet Journal)
Add a temporary constructor for the Station for citadels. This will allow to see Citadel purchases in the Wallet Transaction. Name for citadel is not resolved yet.

Note: should be replaced with a proper citadel system once it's available.
Also, should create an interface for the station & citadel in the WalletTransaction system. (reverted from commit 041c2250bca140e168cc5716484e8b5a59c6978b)

+++ src/EVEMon.Common/Serialization/Eve/SerializableWalletJournalListItem.cs
@@ -34,7 +34,7 @@ namespace EVEMon.Common.Serialization.Eve
         }
 
         [XmlAttribute("argID1")]
-        public long ArgID1 { get; set; }
+        public int ArgID1 { get; set; }
 
         [XmlAttribute("argName1")]
         public string ArgName1 { get; set; }
